refactor(227): new name and place for the ~Service~ type
* type `ServiceOperation` refactored as `OpenAPIService.ts`
* type `ServiceOperation` moved to `OpenAPIService.ts`
