+++ packages/tanstack-query-react-plugin/src/ts-factory/tsdoc/createUseSuspenseInfiniteQueryOperationTSDocExample.ts
@@ -1,4 +1,4 @@
-import { ServiceOperation } from '@openapi-qraft/plugin/lib/open-api/getServices';
+import { ServiceOperation } from '@openapi-qraft/plugin/lib/open-api/OpenAPIService';
 import ts from 'typescript';
 import { astToString } from '../astToString.js';
 import { createInfiniteQueryOptionsArgumentExamplePropertyAssignmentNodes } from './createUseInfiniteQueryOperationTSDocExample.js';
