+++ packages/tanstack-query-react-plugin/src/ts-factory/tsdoc/lib/createOperationMethodCallExpressionExampleNode.ts
@@ -1,4 +1,4 @@
-import { ServiceOperation } from '@openapi-qraft/plugin/lib/open-api/getServices';
+import { ServiceOperation } from '@openapi-qraft/plugin/lib/open-api/OpenAPIService';
 import ts from 'typescript';
 
 export const createOperationMethodCallExpressionExampleNode = (
